requirements.txt
amqp==1.4.7
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.21
celery==3.1.19
certifi==2015.9.6.2
colorama==0.3.3
Django==1.8.5
django-celery==3.1.17
djangorestframework==3.2.5
EasyProcess==0.1.9
fake-factory==0.4.2
flower==0.7.3
gunicorn==19.3.0
kombu==3.0.29
librabbitmq==1.6.1
mysql==0.0.1
MySQL-python==1.2.5
psycopg2==2.6.1
pytz==2015.7
PyVirtualDisplay==0.1.5
selenium==2.46.0
tornado==4.2.1
wheel==0.24.0
sh==1.11
xvfbwrapper==0.2.5